Fix admin set quota
* Move set quota API from AuthController to AdminController.
* Add new error message when the user is not a administrator.
* Add new error message when the number of input bigger than total capacity.
* Only the administrator can set quota.
* Update admin and user docs.
