+++ app/Http/routes.php
@@ -24,7 +24,6 @@ Route::group(['prefix' => 'api', 'middleware' => ['cors', 'api']], function () {
             Route::get('checkEmail/{email}', 'AuthController@checkEmail');
             Route::get('checkCephConnected', 'AuthController@checkCephConnected');
             Route::get('getUserQuota/{user}', 'AuthController@getUserQuota');
-            Route::post('setUserQuota', 'AuthController@setUserQuota');
         });
         Route::group(['middleware' => ['jwt.auth']], function () {
             Route::group(['prefix' => 'bucket', 'namespace' => 'Bucket'], function () {
@@ -53,6 +52,7 @@ Route::group(['prefix' => 'api', 'middleware' => ['cors', 'api']], function () {
                 Route::post('reset', 'AdminController@reset');
                 Route::post('role', 'AdminController@update');
                 Route::delete('delete/{email}', 'AdminController@destroy');
+                Route::post('setQuota', 'AdminController@setQuota');
             });
             Route::group(['prefix' => 'user', 'namespace' => 'User'], function () {
                 Route::get('state', 'UserController@state');
