+++ docs/v1/admin.md
@@ -5,6 +5,7 @@
 3. [Reset Password User](#ResetPasswordUser)
 4. [Update Role User](#UpdateRoleUser)
 5. [Delete User](#DeleteUser)
+6. [Set Quota](#SetQuota)
 
 ## 1.<a name="CreateUser">Create User</a>
 
@@ -287,6 +288,93 @@ status code:403
 status code:403
 {
   "message": "The delete user operation failed"
+}
+```
+
+## 17.<a name="SetUserQuota">Set User Quota</a>
+
+<table>
+    <tr>
+        <td style="width:50px">Method</td>
+        <td style="width:400px">URI</td>
+    </tr>
+    <tr>
+        <td style="width:50px">POST</td>
+        <td style="width:400px">/api/v1/auth/setUserQuota</td>
+    </tr>
+</table>
+
+### Input Parameter
+
+<table>
+    <tr>
+        <td style="width:50px">Type</td>
+        <td style="width:150px">Name</td>
+        <td style="width:50px">Require</td>
+        <td style="width:100px">Remark</td>
+    </tr>
+    <tr>
+        <td style="width:50px">Email</td>
+        <td style="width:150px">email</td>
+        <td style="width:50px">✔︎</td>
+        <td style="width:100px"></td>
+    </tr>
+    <tr>
+        <td style="width:50px">Integer</td>
+        <td style="width:150px">max-objects</td>
+        <td style="width:50px">✔︎</td>
+        <td style="width:100px"></td>
+    </tr>
+    <tr>
+        <td style="width:50px">Integer</td>
+        <td style="width:150px">max-size-kb</td>
+        <td style="width:50px">✔︎</td>
+        <td style="width:100px"></td>
+    </tr>
+    <tr>
+        <td style="width:50px">Integer</td>
+        <td style="width:150px">bucket</td>
+        <td style="width:50px">✔︎</td>
+        <td style="width:100px"></td>
+    </tr>
+    <tr>
+        <td style="width:50px">Boolean</td>
+        <td style="width:150px">enabled</td>
+        <td style="width:50px">✔︎</td>
+        <td style="width:100px"></td>
+    </tr>
+</table>
+
+
+### JSON Response
+#### Success
+```
+status code:200
+{
+  "message": "Setting is successful"
+}
+```
+
+#### Error
+```
+status code:403
+{
+  "message": "Permission denied"
+}
+- or -
+- status code:403
+{
+  "message": "Max size is bigger than variable capacity"
+}
+- or -
+status code:403
+{
+  "message": "The user is not exist"
+}
+- or -
+status code:403
+{
+  "message": "Max Size are not allowed"
 }
 ```
